test(api-db): move expected inventory DB tests from api-core
Move expected rack, switch, and power shelf database-focused tests into the
api-db crate, keeping api-core focused on API handler coverage.
