+++ crates/api-db/src/expected_power_shelf.rs
@@ -326,3 +326,6 @@ pub async fn create_missing_from(
 
     Ok(())
 }
+
+#[cfg(test)]
+mod tests;
