+++ crates/api-db/src/expected_switch.rs
@@ -412,3 +412,6 @@ pub async fn create_missing_from(
 
     Ok(())
 }
+
+#[cfg(test)]
+mod tests;
